backend/app/requirements.txt
Django>=4.1
django-cors-headers>=3.13
djangorestframework>=3.12
gunicorn>=20.1.0
psycopg2-binary>=2.9.1
python-dotenv ~= 0.21
requests ~= 2.28
uWSGI>=2.0.20
django-allauth==0.43